Add files via upload
This is my first project using ml ,I used pandas and LogisticRegression from Sklearn ,I got accurecy 84%.
Dataset description: 
XYZ is a large company with around 4000 employees, and experiences an annual attrition rate of 15%. This level of attrition is considered to be detrimental to the company's reputation and operations. As a result, the management has contracted an HR analytics firm to identify the factors that contribute to attrition and to model the probability of attrition using logistic regression. The goal of this case study is to identify the key variables that the company should focus on in order to reduce attrition and improve employee retention. The results of the analysis will be used by the management to make changes to the workplace and retain more employees.